+++ addon/components/bsy-modal.js
@@ -3,5 +3,15 @@ import layout from '../templates/components/bsy-modal';
 
 export default Ember.Component.extend({
   layout,
-  tagName: "div"
+  tagName: "div",
+
+  didRender() {
+    this._super(...arguments);
+    Ember.$('body').addClass('modal-open');
+  },
+
+  willDestroyElement() {
+    this._super(...arguments);
+    Ember.$('body').removeClass('modal-open');
+  }
 });